requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
flask-bcrypt==1.0.1
pytest==7.3.1
pytest-flask==1.2.0
python-dotenv==1.0.0 
flasgger==0.9.7b2
flask-cors==4.0.1
requests==2.28.1